make config required field, move enum to types
