+++ packages/core/src/hooks/hookPlanner.test.ts
@@ -7,8 +7,7 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { HookPlanner } from './hookPlanner.js';
 import type { HookRegistry, HookRegistryEntry } from './hookRegistry.js';
-import { HookEventName, HookType } from './types.js';
-import { ConfigSource } from './hookRegistry.js';
+import { ConfigSource, HookEventName, HookType } from './types.js';
 
 // Mock debugLogger using vi.hoisted
 const mockDebugLogger = vi.hoisted(() => ({
